+++ astronomer/providers/apache/livy/operators/livy.py
@@ -15,6 +15,9 @@ class LivyOperatorAsync(LivyOperator):
     and set `deferrable` param to `True` instead.
     """
 
+    is_deprecated = True
+    post_deprecation_replacement = "from airflow.providers.apache.livy.operators.livy import LivyOperator"
+
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         warnings.warn(
             "This class is deprecated. "