+++ CHANGELOG.rst
@@ -1,6 +1,23 @@
 Changelog
 =========
 
+1.19.4 (2024-08-23)
+-------------------
+
+Docs
+""""
+
+* Change deprecation status to Yes/No instead of icons (#1572)
+
+
+1.19.3 (2024-08-20)
+-------------------
+
+Docs
+""""
+* Show deprecation status of operators and sensors on the docs operator list page (#1568)
+
+
 1.19.2 (2024-07-29)
 -------------------